Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbobjectstore_delete3-expected.txt


PASS IDBObjectStore.delete() - object store's key path is an object attribute
